+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Changelog
 
+* 6/13/2020: Show opt in and missed numbers in rounds - [@dblock](https://github.com/dblock).
 * 4/11/2020: Added unsubscribe - [@dblock](https://github.com/dblock).
 * 7/2/2018: Added stats per round - [@dblock](https://github.com/dblock).
 * 9/25/2017: Configure followup day - [@ashkan18](https://github.com/ashkan18).